requirements.txt
fastapi
tensorflow
numpy
Pillow
uvicorn
python-multipart